+++ examples/use_config/stm32h7/src/keymap.rs
@@ -1,5 +1,5 @@
 //! You can specify the default keymap here, or in `keyboard.toml`
-//! Check the documentation here: https://haobogu.github.io/rmk/configuration.html#keymap-config
+//! Check the documentation here: https://haobogu.github.io/rmk/keyboard_configuration.html#keymap-config
 //!
 // use crate::{COL, NUM_LAYER, ROW};
 // use rmk::action::KeyAction;